third_party/blink/web_tests/html/tabular_data/table_border_invalid-expected.txt

https://bugs.webkit.org/show_bug.cgi?id=102112: There should be a black box below.


PASS getComputedStyle(document.querySelector("table")).borderTopWidth is "1px"
PASS successfullyParsed is true

TEST COMPLETE
